requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
courlan==1.3.1
dateparser==1.2.0
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
htmldate==1.9.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.1
langchain==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langsmith==0.1.114
lxml==4.9.2
multidict==6.0.5
numpy==1.24.4
openai==1.43.0
orjson==3.10.7
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
six==1.16.0
slack_bolt==1.20.1
slack_sdk==3.31.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
tenacity==8.5.0
tiktoken==0.7.0
tld==0.13
tqdm==4.66.5
trafilatura==1.12.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
yarl==1.9.11
